Update LICENSE file
The primary reason for updating this is to fix the Free Software
Foundation's street address, which has not been at Mass Ave for
many years now.

I used the contents of the license files shipped with Perl 5.38 to
construct the new version. A consequence of this is that the GNU GPL
license text is now that of GPL version 1 rather than version 2
(the Perl license is GPL version 1 or later, or Artistic, which is
why Perl itself includes the GPL version 1 license text).